% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgsearch_functions.R
\docType{package}
\name{kgsearch_googleAuthR}
\alias{kgsearch_googleAuthR}
\alias{kgsearch_googleAuthR-package}
\title{Knowledge Graph Search API
Searches the Google Knowledge Graph for entities.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:59:51
filename: /Users/mark/dev/R/autoGoogleAPI/googlekgsearchv1.auto/R/kgsearch_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}
