% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgsearch_objects.R
\name{SearchResponse}
\alias{SearchResponse}
\title{Knowledge Graph Search API Objects 
Knowledge Graph Search API allows developers to search the Google Knowledge Graph for entities.}
\usage{
SearchResponse(context = NULL, type = NULL, itemListElement = NULL)
}
\arguments{
\item{context}{The local context applicable for the response}

\item{type}{The schema type of top-level JSON-LD object, e}

\item{itemListElement}{The item list of search results}
}
\value{
SearchResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:20:16
filename: /Users/mark/dev/R/autoGoogleAPI/googlekgsearchv1.auto/R/kgsearch_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
SearchResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message includes the context and a list of matching results which contain the detail of associated entities.
}

