% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgsearch_objects.R
\name{SearchResponse}
\alias{SearchResponse}
\title{Knowledge Graph Search API Objects 
Searches the Google Knowledge Graph for entities.}
\usage{
SearchResponse(X.context = NULL, itemListElement = NULL, X.type = NULL)
}
\arguments{
\item{X.context}{The local context applicable for the response}

\item{itemListElement}{The item list of search results}

\item{X.type}{The schema type of top-level JSON-LD object, e}
}
\value{
SearchResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:59:51
filename: /Users/mark/dev/R/autoGoogleAPI/googlekgsearchv1.auto/R/kgsearch_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
SearchResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message includes the context and a list of matching resultswhich contain the detail of associated entities.
}
